rework some tests with pgxmock
