+++ internal/pgengine/bootstrap_test.go
@@ -154,87 +154,6 @@ func TestTryLockClientName(t *testing.T) {
 	})
 }
 
-func TestCreateChainFromYamlErrors(t *testing.T) {
-	initmockdb(t)
-	defer mockPool.Close()
-	mockpge := pgengine.NewDB(mockPool, "pgengine_unit_test")
-
-	t.Run("Database error during chain creation", func(t *testing.T) {
-		mockPool.ExpectQuery(`INSERT INTO timetable.chain`).
-			WithArgs(anyArgs(9)...).
-			WillReturnError(fmt.Errorf("simulated DB error"))
-		_, err := mockpge.CreateChainFromYaml(ctx, &pgengine.YamlChain{})
-		assert.Error(t, err)
-		assert.NoError(t, mockPool.ExpectationsWereMet())
-	})
-
-	t.Run("Database error during task creation", func(t *testing.T) {
-		mockPool.ExpectQuery(`INSERT INTO timetable.chain`).
-			WithArgs(anyArgs(9)...).
-			WillReturnRows(pgxmock.NewRows([]string{"chain_id"}).AddRow(1))
-		mockPool.ExpectQuery(`INSERT INTO timetable.task`).
-			WithArgs(anyArgs(10)...).
-			WillReturnError(fmt.Errorf("simulated DB error on task"))
-
-		_, err := mockpge.CreateChainFromYaml(ctx, &pgengine.YamlChain{
-			Chain:    pgengine.Chain{ChainName: "test-chain"},
-			Schedule: "0 0 * * *",
-			Tasks: []pgengine.YamlTask{
-				{ChainTask: pgengine.ChainTask{Command: "SELECT 1", Kind: "SQL"}},
-			},
-		})
-		assert.Error(t, err)
-		assert.NoError(t, mockPool.ExpectationsWereMet())
-	})
-
-	t.Run("Database error during parameter unmarshalling", func(t *testing.T) {
-		mockPool.ExpectQuery(`INSERT INTO timetable.chain`).
-			WithArgs(anyArgs(9)...).
-			WillReturnRows(pgxmock.NewRows([]string{"chain_id"}).AddRow(1))
-		mockPool.ExpectQuery(`INSERT INTO timetable.task`).
-			WithArgs(anyArgs(10)...).
-			WillReturnRows(pgxmock.NewRows([]string{"task_id"}).AddRow(1))
-
-		_, err := mockpge.CreateChainFromYaml(ctx, &pgengine.YamlChain{
-			Chain:    pgengine.Chain{ChainName: "test-chain"},
-			Schedule: "0 0 * * *",
-			Tasks: []pgengine.YamlTask{
-				{
-					ChainTask:  pgengine.ChainTask{Command: "SELECT 1", Kind: "SQL"},
-					Parameters: []any{func() {}}, // functions cannot be marshalled to JSON
-				},
-			},
-		})
-		assert.Error(t, err)
-		assert.NoError(t, mockPool.ExpectationsWereMet())
-	})
-
-	t.Run("Database error during parameter creation", func(t *testing.T) {
-		mockPool.ExpectQuery(`INSERT INTO timetable.chain`).
-			WithArgs(anyArgs(9)...).
-			WillReturnRows(pgxmock.NewRows([]string{"chain_id"}).AddRow(1))
-		mockPool.ExpectQuery(`INSERT INTO timetable.task`).
-			WithArgs(anyArgs(10)...).
-			WillReturnRows(pgxmock.NewRows([]string{"task_id"}).AddRow(1))
-		mockPool.ExpectExec(`INSERT INTO timetable.parameter`).
-			WithArgs(anyArgs(3)...).
-			WillReturnError(fmt.Errorf("simulated DB error on parameter"))
-
-		_, err := mockpge.CreateChainFromYaml(ctx, &pgengine.YamlChain{
-			Chain:    pgengine.Chain{ChainName: "test-chain"},
-			Schedule: "0 0 * * *",
-			Tasks: []pgengine.YamlTask{
-				{
-					ChainTask:  pgengine.ChainTask{Command: "SELECT 1", Kind: "SQL"},
-					Parameters: []any{"foo"},
-				},
-			},
-		})
-		assert.Error(t, err)
-		assert.NoError(t, mockPool.ExpectationsWereMet())
-	})
-}
-
 func TestExecuteFileScript(t *testing.T) {
 	initmockdb(t)
 	defer mockPool.Close()
